refactoring and add recurring switch
