+++ test/widgets/login.dart
@@ -12,7 +12,7 @@ import 'package:eagler/main.dart';
 void main() {
   testWidgets('there should be a login button', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const App());
     await tester.pumpAndSettle();
     // find the login button
     dynamic loginButton = find.widgetWithText(ElevatedButton, 'Login');
